requirements.txt
streamlit>=1.30.0
torch>=2.1.0
open-clip-torch>=2.24.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
gradio>=6.0.0
